% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_virtual_gene_class.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\title{Gene class}
\value{
This documentation describes the structure of the virtual
\code{Gene} class.
}
\description{
A virtual class to represent genes
}
\details{
The validity function ensures that
\itemize{
\item The \code{id} follows the Ensembl transcript format
\code{^ENST[0-9]+$}.
\item The \code{hugo_symbol} slot contains only uppercase letters, digits,
and special characters '-' or '_'.
\item The \code{name}, and \code{description} are non-empty strings,
if specified.
\item The \code{tissue_specificity} slot is a list of tissues,
if specified.
\item The \code{gene_structure} slot is a \code{GRanges} object with
valid chromosome, and strand. The \code{IRanges}
object in \code{GRanges} directly verify that 'start' is less than or
equal to 'end' when the user creates the GRanges object, so this check is
not implemented in the package.
\item The \code{gene_product} slot is a list that represents a gene
product and contains two elements: an ID and a sequence.
\item The \code{clinical_significance} is a non-empty string,
if specified.
}
}
\section{Slots}{

\describe{
\item{\code{id}}{character. Ensembl transcript ID of the gene. In this
implementation, each transcript isoform is considered as an individual gene.
Hence, the Ensembl transcript ID ("ENST") is used instead of the Ensembl
gene ID ("ENSG"). This approach simplifies the management of Gene objects,
avoiding the complexity of handling multiple gene products of different
types within a single Gene object.}

\item{\code{hugo_symbol}}{character. Hugo symbol of the gene.}

\item{\code{name}}{character. Optional parameter. Name of the gene.}

\item{\code{description}}{character. Optional parameter.
Description of the gene.}

\item{\code{tissue_specificity}}{list. Optional parameter. List of tissues
(character) where the gene is specifically expressed.}

\item{\code{gene_structure}}{GenomicRanges::GRanges. Structure of the gene,
including chromosomes ("seqnames", character), start position
("start", numeric), end position ("end", numeric),
strand ("strand", character).}

\item{\code{gene_product}}{list. It represents the product of the gene and
contains its product id (character) and the corresponding
sequence (character).}

\item{\code{clinical_significance}}{character. Optional parameter. Clinical
relevance of the gene or its product (e.g., association with a disease
or therapeutic targets).}
}}

\note{
The \code{gene_structure} and \code{gene_product} arguments are
complex objects and should be created separately by the user before
constructing the Gene object. This approach maximizes flexibility, allowing
the user to better exploit the functionalities provided by the
\code{GenomicRanges} package and to easily reuse the \code{gene_structure}
object in other contexts.
}
\keyword{classes}
