% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_housekeepingRNAgene_classes.R
\docType{class}
\name{HousekeepingRNAGene-class}
\alias{HousekeepingRNAGene-class}
\title{Housekeeping RNA gene class}
\value{
This documentation describes the structure of the virtual
\code{HousekeepingRNAGene} class.
}
\description{
A virtual class to represent housekeeping RNA genes.
This class is a general S4 class inheriting from the \code{Gene}
class and serves as a base class to represent specific types of
housekeeping RNA gene.
}
\keyword{internal}
