% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_lengthProduct_function.R
\name{lengthProduct}
\alias{lengthProduct}
\alias{lengthProduct,Gene-method}
\title{Compute the length of a gene product for a gene object}
\usage{
lengthProduct(object)

\S4method{lengthProduct}{Gene}(object)
}
\arguments{
\item{object}{Gene object. An object of a specific Gene class
(e.g., "ProteinCodingGene", "LncRNAGene", "SiRNAGene", etc.).}
}
\value{
an integer representing the length of the gene product.
}
\description{
This function computes the length of the product of a gene object.
It provides an integer value that represents the number of nucleotides
or amino acids in the sequence, depending on the Gene object class.
}
\details{
Based on the class of the Gene object, the function choose the
appropriate method to compute the length of the gene product.
}
\examples{
gene_structure <- GenomicRanges::GRanges(seqnames = "chr1",
                  ranges = IRanges::IRanges(start = 200, end = 1200),
                  strand = "+")
                         
gene_product <- list(lncrna_id = "lncRNAID", 
                     lncrna_sequence = paste0(
                       "AUGCUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCG",
                       "GCUAAUCGGCUUAGCGUACGGUAGGCUUAACUGCGUACGAU",
                       "CGAUCGGCUAAGCGUACGGUAGGCUUAACUGCGUACGAUCG",
                       "AUCGGCUUAGCGUACGUAGGCUCAACUGCGUACGAUCGAUC",
                       "GGCUUAGCGUACGGUAGGCUUAACUGCGGACGAUCGAUCGG",
                       "CUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCGC"))
                                          
lncrna_gene <- createLncRNAGene(id = "ENST000001",
               hugo_symbol = "SYMBOL1",
               name = "lncRNA gene name",
               description = "gene description",
               tissue_specificity = list("liver", "small bowel"),
               gene_structure = gene_structure,
               gene_product = gene_product,
               clinical_significance = "association with disease")

lengthProduct(lncrna_gene)

}
\keyword{gene,}
\keyword{length,}
\keyword{product}
\keyword{sequence,}
