% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_lncRNAgene_class.R
\docType{class}
\name{LncRNAGene-class}
\alias{LncRNAGene-class}
\title{Long non-coding RNA Gene class}
\value{
This documentation describes the structure of the
\code{LncRNAGene} class.
}
\description{
A class to represent long non-coding RNA genes
This class is a specialized S4 class inheriting from the \code{Gene}
class and is designed to store information about long non-coding RNA
genes, including their lncRNA product.
}
\details{
The \code{gene_product} slot is expected to contain a lncRNA ID and the
corresponding sequence. The validity function ensures that it is correctly
formatted and contain a valid lncRNA sequence.
}
\section{Slots}{

\describe{
\item{\code{gene_product}}{list. This slot is specific for the long non-coding
RNA genes product and includes:
\itemize{
\item \code{lncrna_id}: a string representing the ID of the lncRNA.
\item \code{lncrna_sequence}: a string representing the sequence of
the lncRNA.
}}
}}

