% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_sncRNAgene_classes.R
\docType{class}
\name{MicroRNAGene-class}
\alias{MicroRNAGene-class}
\title{MicroRNA gene class}
\value{
This documentation describes the structure of the
\code{MicroRNAGene} class.
}
\description{
A class to represent microRNA genes.
This class is a specialized S4 class inheriting from the \code{Gene}
class and is designed to store information about microRNA genes,
including their microRNA product.
}
\details{
The \code{gene_product} slot is expected to contain a microRNA ID and
the corresponding sequence. The validity function ensures that it is
correctly formatted and contain a valid microRNA sequence.
}
\section{Slots}{

\describe{
\item{\code{gene_product}}{list. This slot is specific for the microRNA genes
product and includes:
\itemize{
\item \code{microrna_id}: a string representing the ID of the microRNA.
\item \code{microrna_sequence}: a string representing the sequence of
the microRNA.
}}
}}

