% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_sncRNAgene_classes.R
\docType{class}
\name{PiRNAGene-class}
\alias{PiRNAGene-class}
\title{Piwi-interacting RNA gene class}
\value{
This documentation describes the structure of the
\code{PiRNAGene} class.
}
\description{
A class to represent piwi-interacting RNA genes.
This class is a specialized S4 class inheriting from the \code{Gene}
class and is designed to store information about piRNA genes, including
their piRNA products.
}
\details{
The \code{gene_product} slot is expected to contain a piRNA ID and the
corresponding sequence. The validity function ensures that it is correctly
formatted and contain a valid piRNA sequence.
}
\section{Slots}{

\describe{
\item{\code{gene_product}}{list. This slot is specific for the piRNA genes
product and includes:
\itemize{
\item \code{pirna_id}: a string representing the ID of the piRNA.
\item \code{pirna_sequence}: a string representing the sequence of
the piRNA.
}}
}}

