% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coding_RNAgene_class.R
\docType{class}
\name{ProteinCodingGene-class}
\alias{ProteinCodingGene-class}
\title{Protein-coding RNA gene class}
\value{
This documentation describes the structure of the
\code{ProteinCodingGene} class.
}
\description{
A class to represent protein-coding genes.
This class is a specialized S4 class inheriting from the \code{Gene}
class and is designed to store information about protein-coding
genes, including their protein product.
}
\details{
The \code{gene_product} slot is expected to contain a protein ID and
the corresponding sequence. The validity function ensures that it is
correctly formatted and contain a valid protein sequence.
}
\section{Slots}{

\describe{
\item{\code{gene_product}}{list. This slot is specific for the
protein-coding genes product and includes:
\itemize{
\item \code{protein_id}: a string representing the ID of the protein.
\item \code{protein_sequence}: a string representing the sequence of
the protein.
}}
}}

