% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_get_set_Ensembl_id.R
\name{setID<-}
\alias{setID<-}
\alias{setID<-,Gene-method}
\title{Set the Ensembl transcript ID}
\usage{
setID(object) <- value

\S4method{setID}{Gene}(object) <- value
}
\arguments{
\item{object}{Gene object.}

\item{value}{the new Ensembl transcript ID.}
}
\value{
the modified Gene object.
}
\description{
This function sets the Ensembl transcript ID for a Gene object.
}
\details{
This function is defined as a generic function to be applicable
to any object that inherits from the Gene class. After setting the new ID,
the function checks that the Gene object is still valid
by calling \code{validObject}.
}
\examples{
gene_structure <- GenomicRanges::GRanges(seqnames = "chr1",
                  ranges = IRanges::IRanges(start = 200, end = 1200),
                  strand = "+")
                         
gene_product <- list(lncrna_id = "lncRNAID", 
                     lncrna_sequence = paste0(
                       "AUGCUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCG",
                       "GCUAAUCGGCUUAGCGUACGGUAGGCUUAACUGCGUACGAU",
                       "CGAUCGGCUAAGCGUACGGUAGGCUUAACUGCGUACGAUCG",
                       "AUCGGCUUAGCGUACGUAGGCUCAACUGCGUACGAUCGAUC",
                       "GGCUUAGCGUACGGUAGGCUUAACUGCGGACGAUCGAUCGG",
                       "CUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCGC"))
                                          
lncrna_gene <- createLncRNAGene(id = "ENST000001",
               hugo_symbol = "SYMBOL1",
               name = "lncRNA gene name",
               description = "gene description",
               tissue_specificity = list("liver", "small bowel"),
               gene_structure = gene_structure,
               gene_product = gene_product,
               clinical_significance = "association with disease")
                                 
setID(lncrna_gene) <- "ENST000002"

getID(lncrna_gene)                                 

}
