% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_get_set_gene_product.R
\name{setProductSequence<-}
\alias{setProductSequence<-}
\alias{setProductSequence<-,Gene-method}
\title{Set the gene product sequence}
\usage{
setProductSequence(object) <- value

\S4method{setProductSequence}{Gene}(object) <- value
}
\arguments{
\item{object}{Gene object.}

\item{value}{the new product sequence of the gene product.}
}
\value{
the modified Gene object.
}
\description{
This function sets the gene product sequence.
}
\details{
This function is defined as a generic function to be applicable
to any object that inherits from the Gene class. After setting the new
product sequence, the function checks that the Gene object
is still valid by calling \code{validObject}.
}
\examples{
gene_structure <- GenomicRanges::GRanges(seqnames = "chr1",
                  ranges = IRanges::IRanges(start = 200, end = 1200),
                  strand = "+")
                         
gene_product <- list(lncrna_id = "lncRNAID", 
                     lncrna_sequence = paste0(
                       "AUGCUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCG",
                       "GCUAAUCGGCUUAGCGUACGGUAGGCUUAACUGCGUACGAU",
                       "CGAUCGGCUAAGCGUACGGUAGGCUUAACUGCGUACGAUCG",
                       "AUCGGCUUAGCGUACGUAGGCUCAACUGCGUACGAUCGAUC",
                       "GGCUUAGCGUACGGUAGGCUUAACUGCGGACGAUCGAUCGG",
                       "CUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCGC"))
                                          
lncrna_gene <- createLncRNAGene(id = "ENST000001",
               hugo_symbol = "SYMBOL1",
               name = "lncRNA gene name",
               description = "gene description",
               tissue_specificity = list("liver", "small bowel"),
               gene_structure = gene_structure,
               gene_product = gene_product,
               clinical_significance = "association with disease")
                                 
setProductSequence(lncrna_gene) <- paste0(
                     "AUGCUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCG",
                     "CGAUCGGCUAAGCGUACGGUAGGCUUAACUGCGUACGAUCG",
                     "GCUAAUCGGCUUAGCGUACGGUAGGCUUAACUGCGUACGAU",
                     "GGCUUAGCGUACGGUAGGCUUAACUGCGGACGAUCGAUCGG",
                     "AUCGGCUUAGCGUACGUAGGCUCAACUGCGUACGAUCGAUC",
                     "CUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCGC")

getProductSequence(lncrna_gene)    

}
