% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_showGeneObject_function.R
\name{showGeneObject}
\alias{showGeneObject}
\alias{showGeneObject,Gene-method}
\title{Show Gene object information}
\usage{
showGeneObject(object)

\S4method{showGeneObject}{Gene}(object)
}
\arguments{
\item{object}{Gene object. An object of a specific Gene class
(e.g., "ProteinCodingGene", "LncRNAGene", "SiRNAGene", etc.).}
}
\value{
This function does not return a value. It displays the gene object
information to the console.
}
\description{
This function displays the gene object information in a more suitable
and detailed way with respect to the standard 'show' method
implemented in R.
}
\details{
The gene_structure slot is an S4 object of class GRanges, which
is implemented in order to give more flexibility in its creation.
To ensure a more standardized and reliable output format, the R \code{show}
function is used within the \code{showGeneObject} function to display it.
Additionally, the product sequence is formmatted into blocks, with each
line containing 80 bases for better readability.
}
\examples{
gene_structure <- GenomicRanges::GRanges(seqnames = "chr1",
                  ranges = IRanges::IRanges(start = 200, end = 1200),
                  strand = "+")
                         
gene_product <- list(lncrna_id = "lncRNAID", 
                     lncrna_sequence = paste0(
                       "AUGCUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCG",
                       "GCUAAUCGGCUUAGCGUACGGUAGGCUUAACUGCGUACGAU",
                       "CGAUCGGCUAAGCGUACGGUAGGCUUAACUGCGUACGAUCG",
                       "AUCGGCUUAGCGUACGUAGGCUCAACUGCGUACGAUCGAUC",
                       "GGCUUAGCGUACGGUAGGCUUAACUGCGGACGAUCGAUCGG",
                       "CUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCGC"))
                                          
lncrna_gene <- createLncRNAGene(id = "ENST000001",
               hugo_symbol = "SYMBOL1",
               name = "lncRNA gene name",
               description = "gene description",
               tissue_specificity = list("liver", "small bowel"),
               gene_structure = gene_structure,
               gene_product = gene_product,
               clinical_significance = "association with disease")

showGeneObject(lncrna_gene)

}
