% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_sncRNAgene_classes.R
\docType{class}
\name{SiRNAGene-class}
\alias{SiRNAGene-class}
\title{Short-interfering RNA gene class}
\value{
This documentation describes the structure of the
\code{SiRNAGene} class.
}
\description{
A class to represent short-interfering RNA genes.
This class is a specialized S4 class inheriting from the \code{Gene}
class and is designed to store information about siRNA genes,
including their siRNA product.
}
\details{
The \code{gene_product} slot is expected to contain a siRNA ID and the
corresponding sequence. The validity function ensures that it is correctly
formatted and contain a valid siRNA sequence.
}
\section{Slots}{

\describe{
\item{\code{gene_product}}{list. This slot is specific for the siRNA genes
product and includes:
\itemize{
\item \code{sirna_id}: a string representing the ID of the siRNA.
\item \code{sirna_sequence}: a string representing the sequence of
the siRNA.
}}
}}

