% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_sncRNAgene_classes.R
\docType{class}
\name{SncRNAGene-class}
\alias{SncRNAGene-class}
\title{Small non-coding RNA gene class}
\value{
This documentation describes the structure of the virtual
\code{SncRNAGene} class.
}
\description{
A virtual class to represent small non-coding RNA genes.
This class is a general S4 class inheriting from the \code{Gene} class
and serves as a base class to represent specific types of non small
non-coding RNA gene.
}
\keyword{internal}
