% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_sncRNAgene_classes.R
\docType{class}
\name{SnoRNAGene-class}
\alias{SnoRNAGene-class}
\title{Small nucleolar RNA gene class}
\value{
This documentation describes the structure of the
\code{SnoRNAGene} class.
}
\description{
A class to represent small nucleolar RNA genes.
This class is a specialized S4 class inheriting from the \code{Gene}
class and is designed to store information about snoRNA genes,
including their snoRNA products.
}
\details{
The \code{gene_product} slot is expected to contain a snoRNA ID and
the corresponding sequence. The validity function ensures that it is
correctly formatted and contain a valid snoRNA sequence.
}
\section{Slots}{

\describe{
\item{\code{gene_product}}{list. This slot is specific for the snoRNA genes
product and includes:
\itemize{
\item \code{snorna_id}: a string representing the ID of the snoRNA.
\item \code{snorna_sequence}: a string representing the sequence of
the snoRNA.
}}
}}

