% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_sncRNAgene_classes.R
\docType{class}
\name{SnRNAGene-class}
\alias{SnRNAGene-class}
\title{Small nuclear RNA gene class}
\value{
This documentation describes the structure of the
\code{SnRNAGene} class.
}
\description{
A class to represent small nuclear RNA genes.
This class is a specialized S4 class inheriting from the \code{Gene}
class and is designed to store information about snRNA genes,
including their snRNA products.
}
\details{
The \code{gene_product} slot is expected to contain a snRNA ID and
the corresponding sequence. The validity function ensures that it is
correctly formatted and contain a valid snRNA sequence.
}
\section{Slots}{

\describe{
\item{\code{gene_product}}{list. This slot is specific for the snRNA genes
product and includes:
\itemize{
\item \code{snrna_id}: a string representing the ID of the snRNA.
\item \code{snrna_sequence}: a string representing the sequence of
the snRNA.
}}
}}

