% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_housekeepingRNAgene_classes.R
\docType{class}
\name{TRNAGene-class}
\alias{TRNAGene-class}
\title{Transfer RNA gene class}
\value{
This documentation describes the structure of the
\code{TRNAGene} class.
}
\description{
A class to represent transfer RNA genes.
This class is a specialized S4 class inheriting from the \code{Gene}
class and is designed to store information about tRNA genes,
including their tRNA products.
}
\details{
The \code{gene_product} slot is expected to contain a tRNA ID and the
corresponding sequence. The validity function ensures that it is correctly
formatted and contain a valid tRNA sequence.
}
\section{Slots}{

\describe{
\item{\code{gene_product}}{list. This slot is specific for the tRNA genes
product and includes:
\itemize{
\item \code{trna_id}: a string representing the ID of the tRNA.
\item \code{trna_sequence}: a string representing the sequence of
the tRNA.
}}
}}

