% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_validation_function_gene_product.R
\name{validateGeneProduct}
\alias{validateGeneProduct}
\title{Validate Gene Product}
\usage{
validateGeneProduct(id, sequence, type)
}
\arguments{
\item{id}{character. The gene product ID to validate.}

\item{sequence}{character. The gene product sequence to validate.}

\item{type}{character. The RNA type: "lncRNA", "sncRNA" (including
microRNA, piRNA, snRNA, snoRNA, siRNA), "tRNA",
"rRNA", "protein".}
}
\value{
TRUE if both the ID and sequence are valid, throws an error
otherwise.
}
\description{
This function validates both the gene product ID and sequence
using internal methods. Users can use it to validate the
gene products before creating the gene object. However, the function
is also internal in the definition of the gene object classes to
guarantee robustness.
}
\examples{
lncrna_id <- "lncRNAID"

lncrna_sequence <- paste0(
                   "AUGCUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCG",
                   "GCUAAUCGGCUUAGCGUACGGUAGGCUUAACUGCGUACGAU",
                   "CGAUCGGCUAAGCGUACGGUAGGCUUAACUGCGUACGAUCG",
                   "AUCGGCUUAGCGUACGUAGGCUCAACUGCGUACGAUCGAUC",
                   "GGCUUAGCGUACGGUAGGCUUAACUGCGGACGAUCGAUCGG",
                   "CUUAGCGUACGGUAGGCUUAACUGCGUACGAUCGAUCGC") 
type <- "lncRNA"

validateGeneProduct(lncrna_id, lncrna_sequence, type)


}
\keyword{function}
\keyword{gene}
\keyword{product}
\keyword{validation}
