% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_validation_function_gene_product.R
\name{validateID}
\alias{validateID}
\title{Validate Gene Product ID}
\usage{
validateID(id)
}
\arguments{
\item{id}{character. The gene product ID to validate.}
}
\value{
TRUE if the gene product ID is valid, throws an error otherwise.
}
\description{
This function validates the gene product ID to ensure it is a
non-empty string.
}
\keyword{internal}
