% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_validation_function_gene_product.R
\name{validateSequence}
\alias{validateSequence}
\title{Validate Gene Product Sequence}
\usage{
validateSequence(sequence, type)
}
\arguments{
\item{sequence}{character. The gene product sequence to validate.}

\item{type}{character. The gene product type (e.g., "lncRNA", "sncRNA",
"tRNA").}
}
\value{
TRUE if the gene product sequence is valid, throws an error
otherwise.

TRUE if the gene product length is valid, throws an error otherwise.
}
\description{
This function validates the gene product sequence, ensuring it is a
string containing only valid nucleotide bases or amino acids,
depending on the Gene object class. Besides, it checks if the gene
product sequence meets specific length requirements based on its type.
}
\details{
Regarding the length requirements, the function allows some
flexibility due to the variable length ranges defined for some gene types.
A long non-coding RNA gene product must be at least 200 nucleotides
in length. All small non coding genes are considered together and their
product must be shorter than 200 nucleotides. A tRNA gene product ranges
from 70 to 90 nucleotides, in general. No checks are defined for rRNA gene
products because of their high variability in nucleotide sequences instead.
}
\keyword{internal}
